\name{lsmapps_navbar}
\alias{lsmapps_navbar}
\title{Navigation bar and dropdown menus for all polmeth pages}
\usage{
  lsmapps_navbar()
}
\value{
  a tag object that shiny understands
}
\description{
  Creates HTML code which contains the navigation bar and
  dropdown menus common to all lsmapps pages.
}
\examples{
# view generated html code
as.character(lsmapps_navbar())
}

